+++ src/config/yaml_decoders.cc
@@ -468,6 +468,18 @@ template <> struct convert<layout_hint> {
             rhs.hint = hint_t::right;
             rhs.entity = node["right"].as<std::string>();
         }
+        else if (node["together"]) {
+            rhs.hint = hint_t::together;
+            rhs.entity = node["together"].as<std::vector<std::string>>();
+        }
+        else if (node["row"]) {
+            rhs.hint = hint_t::row;
+            rhs.entity = node["row"].as<std::vector<std::string>>();
+        }
+        else if (node["column"]) {
+            rhs.hint = hint_t::column;
+            rhs.entity = node["column"].as<std::vector<std::string>>();
+        }
         else
             return false;
 
@@ -645,7 +657,8 @@ config load(
             doc["git"] = git_config;
         }
 
-        return doc.as<config>();
+        auto d = doc.as<config>();
+        return d;
     }
     catch (YAML::BadFile &e) {
         throw std::runtime_error(fmt::format(