+++ tests/t00053/t00053.cc
@@ -1,0 +1,38 @@
+namespace clanguml {
+namespace t00053 {
+struct a {
+};
+struct b {
+};
+struct c {
+};
+struct d {
+};
+struct e {
+};
+struct f {
+};
+struct g {
+};
+
+struct A {
+};
+struct B {
+};
+struct C {
+};
+struct D {
+};
+struct E {
+};
+struct F {
+};
+struct G {
+};
+
+enum class h { hhh };
+enum class i { iii };
+enum class j { jjj };
+
+}
+}