+++ tests/t00054/t00054.cc
@@ -1,0 +1,48 @@
+namespace clanguml {
+namespace t00054 {
+struct a {
+};
+struct b {
+};
+
+namespace detail {
+struct c {
+};
+struct d {
+};
+struct e {
+};
+} // namespace detail
+struct f {
+};
+struct g {
+};
+
+struct A {
+};
+struct B {
+};
+
+namespace detail2 {
+struct C {
+};
+namespace detail3 {
+struct D {
+};
+struct E {
+};
+} // namespace detail3
+struct F {
+};
+} // namespace detail2
+struct G {
+};
+
+namespace detail4 {
+enum class h { hhh };
+enum class i { iii };
+enum class j { jjj };
+} // namespace detail4
+
+}
+}